Reduce logging when using @! interactively
if someone is using @! interactively they're probably not interested in
tight resource management so reduce warnings to debug level in this
case.

Heuristically, we take "interactive" to be use from the REPL or the Main
module at global scope.

Also fix a bug which swapped file and line info in logging.