Remove @! run()
It's not clear that calling `wait` on the result of `run` has the
correct semantics - unlike other resources, the cleanup of the process
may require nontrivial bespoke communication with the caller.

So let's just remove this for now, perhaps in the future we can have a
more complete story for concurrency in this package.