Implement enter_do in terms of throwto
For this specialized use case in enter_do we really only want the use of
an extra stack, but other than this, we probably don't need to involve
the main scheduler or have any other unrelated tasks be scheduled.

So it seems we can skip the main scheduler completely on the happy path.
When there's an error we do need the child task to run to completion so
that its error flag is set.